Changes to the txData object to update the UID keys used to identify nodes and relationships.
The parameters for the trigger controlling the UID generated in the txData are 'uidKeys' and uidLabels'. If a node does not have a label in the uidLabels list then it will use its first label. If a node/rel does not have a property in uidKeys then it will automatically use the node/rel id.